redis index problem fixed

+++ src/main/kotlin/com/range/killerbot/service/MessageSaveService.kt
@@ -4,6 +4,6 @@ import com.range.killerbot.domain.entity.MessageModel
 
 interface MessageSaveService {
     fun save(message: String,author: String): MessageModel
-    fun countUserId(author: String): Int
+    fun countUserId(author: String): Long
     fun deleteUserMessages(author: String)
 }
